pyproject.toml
[project]
name = "course-management-system"
version = "0.1.0"
description = ""
authors = [
    {name = "officialalkenes",email = "belloabdulhakeemolamide@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "celery (>=5.5.2,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
django-debug-toolbar = "^5.2.0"
django-extensions = "^4.1"
flake8 = "^7.2.0"
ruff = "^0.11.10"
isort = "^6.0.1"
pytest-django = "^4.11.1"
faker = "^37.3.0"


[tool.poetry.group.staging.dependencies]
sentry-sdk = "^2.28.0"


[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.dev"
python_files = "test_*.py tests.py"
python_functions = "test_*"
addopts = "-ra"
testpaths = [
    "core/apps"
]
